backend/requirements.txt
flask==2.3.3
mysql-connector-python==8.0.33
Werkzeug==2.3.7
qrcode==7.4.2
Pillow==10.0.0
reportlab==4.0.7
python-dotenv==1.0.0